requirements.txt
opencv-python 
dlib
face_recognition
face_recognition_models
replicate 
python-dotenv
pydrive